Move/extract CORECONFDatabase into its own module
